IDDL 문서화: 34개 문서로 구성된 완전한 학습 커리큘럼 추가
완전한 6단계 프로그레시브 학습 경로 (Level 0-5 + 부록)

**문서 구성:**
- Level 0: 시작하기 (3개 문서, 30분)
- Level 1: 핵심 개념 (5개 문서, 1시간)
- Level 2: 구조 이해하기 (5개 문서, 1시간)
- Level 3: 데이터와 상호작용 (5개 문서, 1.5시간)
- Level 4: 실전 패턴 (5개 문서, 2시간)
- Level 5: 고급 주제 (5개 문서, 2시간)
- 부록: 참조 문서 (5개 문서)

**문서 특징:**
- 입문부터 마스터까지 점진적 난이도 상승
- 모든 개념마다 실행 가능한 JSON 예제 제공
- 접을 수 있는 인터랙티브 연습문제
- 흔한 실수 섹션 (❌ 잘못된 예 vs ✅ 올바른 예)
- 빠른 참조를 위한 핵심 정리
- 문서 간 이동을 위한 네비게이션 링크

**README.md 재작성:**
- Why, Background, Core Concepts 중심으로 재구성
- 의도 중심 설계 철학 설명
- IDDL 생태계 아키텍처 다이어그램
- Before/After 비교 예제 추가
- 완전한 34개 문서 커리큘럼 링크

**IDDL 명세 추가:**
- spec/iddl-spec-1.0.0.md: 초기 명세
- spec/iddl-spec-1.0.1.md: 업데이트된 명세
- spec/iddl-coverage-analysis.md: 커버리지 분석
