+++ apps/docs/00-getting-started/02-core-idea.md
@@ -1,0 +1,443 @@
+# IDDL의 핵심 아이디어 💡
+
+**예상 소요 시간**: 8분
+**난이도**: ⭐ 입문
+**사전 지식**: [첫 번째 UI 만들기](./01-your-first-ui.md)
+
+---
+
+## 이 문서를 읽고 나면
+
+- "의도 선언 vs 구현 지시"의 차이를 이해합니다
+- 4가지 질문으로 모든 UI를 정의할 수 있습니다
+- IDDL이 왜 강력한지 알 수 있습니다
+
+---
+
+## 핵심 철학: 의도 선언 vs 구현 지시
+
+### 기존 방식: 구현을 지시한다
+
+```css
+/* CSS: 어떻게 그릴지 지시 */
+.delete-button {
+  background-color: #E53E3E;
+  color: white;
+  padding: 12px 24px;
+  border-radius: 6px;
+  font-weight: 600;
+  box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+}
+```
+
+**문제**:
+- ❌ "왜 빨간색인지" 알 수 없음
+- ❌ 다크모드에서 어떤 색으로 바뀔지 모름
+- ❌ AI가 이해할 수 없음 (숫자 더미일 뿐)
+- ❌ 6개월 후 유지보수 불가능
+
+### IDDL 방식: 의도를 선언한다
+
+```json
+{
+  "type": "Action",
+  "label": "Delete",
+  "prominence": "Primary",
+  "intent": "Critical"
+}
+```
+
+**장점**:
+- ✅ "위험한 액션"이라는 의도가 명확함
+- ✅ 디자인 시스템이 맥락에 맞게 렌더링
+- ✅ AI가 이해하고 생성 가능
+- ✅ 영원히 의미가 유지됨
+
+---
+
+## 4가지 질문으로 정의하는 UI
+
+IDDL은 모든 UI 요소를 **4가지 질문**으로 정의합니다:
+
+### 1️⃣ 무엇인가? (`role`)
+
+"이것의 **기능적 정체성**은 무엇인가?"
+
+```json
+// 제목
+{ "role": "Title" }
+
+// 데이터 입력 폼
+{ "role": "Form" }
+
+// 클릭 가능한 버튼
+{ "role": "Control" }
+
+// 카드 컨테이너
+{ "role": "Card" }
+```
+
+### 2️⃣ 얼마나 눈에 띄어야 하는가? (`prominence`)
+
+"시각적 계층에서 **몇 번째**로 중요한가?"
+
+```json
+// 페이지 메인 타이틀 (가장 크고 굵게)
+{ "prominence": "Hero" }
+
+// 주요 버튼 (표준 크기)
+{ "prominence": "Primary" }
+
+// 보조 텍스트 (작고 흐리게)
+{ "prominence": "Secondary" }
+
+// 메타 정보 (거의 안 보이게)
+{ "prominence": "Tertiary" }
+```
+
+### 3️⃣ 무슨 맥락(색상)인가? (`intent`)
+
+"이것이 전달하는 **의미**는 무엇인가?"
+
+```json
+// 일반 정보 (회색)
+{ "intent": "Neutral" }
+
+// 브랜드 강조 (Primary 색상)
+{ "intent": "Brand" }
+
+// 성공/긍정 (녹색)
+{ "intent": "Positive" }
+
+// 주의/경고 (노란색)
+{ "intent": "Caution" }
+
+// 위험/삭제 (빨간색)
+{ "intent": "Critical" }
+```
+
+### 4️⃣ 얼마나 빽빽하게 보여줄 것인가? (`density`)
+
+"정보를 **어떤 간격**으로 배치할 것인가?"
+
+```json
+// 마케팅 페이지, 대시보드 (넓은 여백)
+{ "density": "Comfortable" }
+
+// 일반 문서, 폼 (표준 간격)
+{ "density": "Standard" }
+
+// 데이터 그리드, 전문가 도구 (좁은 간격)
+{ "density": "Compact" }
+```
+
+---
+
+## 실전 비교: 버튼 3종
+
+같은 버튼도 **의도에 따라** 다르게 표현됩니다.
+
+### 1. 메인 CTA (브랜드 강조)
+
+```json
+{
+  "type": "Action",
+  "label": "Get Started",
+  "prominence": "Hero",
+  "intent": "Brand"
+}
+```
+
+→ **가장 크고 브랜드 색상**으로 렌더링 (Primary 버튼)
+
+### 2. 저장 버튼 (성공 액션)
+
+```json
+{
+  "type": "Action",
+  "label": "Save Changes",
+  "prominence": "Primary",
+  "intent": "Positive"
+}
+```
+
+→ **표준 크기, 녹색** (성공 버튼)
+
+### 3. 삭제 버튼 (위험 액션)
+
+```json
+{
+  "type": "Action",
+  "label": "Delete",
+  "prominence": "Secondary",
+  "intent": "Critical",
+  "confirm": "Are you sure?"
+}
+```
+
+→ **작은 크기, 빨간색, 확인 메시지** (위험 버튼)
+
+---
+
+## 왜 이 방식이 강력한가?
+
+### 1. AI와의 완벽한 협업
+
+```
+User: "사용자 삭제 버튼 만들어줘, 경고 메시지 포함"
+
+GPT:
+{
+  "type": "Action",
+  "label": "Delete User",
+  "prominence": "Secondary",
+  "intent": "Critical",
+  "confirm": "This action cannot be undone",
+  "behavior": {
+    "action": "command",
+    "command": "user.delete"
+  }
+}
+```
+
+AI가 **의도를 이해**하고 적절한 속성을 자동 선택합니다.
+
+### 2. 디자인 시스템 독립적
+
+동일한 IDDL이 다양한 스타일로 렌더링됩니다:
+
+```json
+{
+  "type": "Action",
+  "prominence": "Primary",
+  "intent": "Brand",
+  "label": "Submit"
+}
+```
+
+**Material Design**: 채워진 파란색 버튼
+**iOS Style**: 둥근 모서리, 그라데이션
+**Windows 11**: 각진 모서리, Fluent Design
+**CLI**: `[Submit] ←`
+
+### 3. 유지보수성
+
+```json
+// 6개월 후에도 의미가 명확
+{
+  "type": "Action",
+  "intent": "Critical",  // 아, 위험한 액션이구나
+  "confirm": "..."        // 확인 메시지 있구나
+}
+```
+
+vs
+
+```css
+/* 6개월 후 ??? */
+.btn-red {
+  background: #E53E3E;  // 왜 빨간색이지?
+}
+```
+
+### 4. 확장성
+
+새로운 요구사항이 생겨도 **의도는 그대로**:
+
+```json
+// 다크모드 추가? → 디자인 시스템이 자동 처리
+// 모바일 대응? → 디자인 시스템이 자동 처리
+// 접근성? → role과 prominence로 자동 생성
+```
+
+---
+
+## 실습: 의도를 찾아라
+
+다음 UI를 보고 적절한 IDDL 속성을 선택해보세요.
+
+### 문제 1: 페이지 제목
+
+```
+━━━━━━━━━━━━━━━━━━━━
+Welcome to Dashboard
+━━━━━━━━━━━━━━━━━━━━
+```
+
+<details>
+<summary>정답 보기</summary>
+
+```json
+{
+  "type": "Text",
+  "role": "Title",
+  "content": "Welcome to Dashboard",
+  "prominence": "Hero"  // 페이지 최상위 제목
+}
+```
+
+</details>
+
+### 문제 2: 비활성 계정 알림
+
+```
+┌────────────────────────┐
+│ ⚠ Your account is      │
+│ inactive. Please       │
+│ verify your email.     │
+│ [Resend Email]         │
+└────────────────────────┘
+```
+
+<details>
+<summary>정답 보기</summary>
+
+```json
+{
+  "type": "Group",
+  "role": "Card",
+  "intent": "Caution",  // 경고 맥락
+  "children": [
+    {
+      "type": "Text",
+      "role": "Body",
+      "content": "Your account is inactive. Please verify your email.",
+      "prominence": "Primary"
+    },
+    {
+      "type": "Action",
+      "label": "Resend Email",
+      "prominence": "Secondary",  // 보조 액션
+      "intent": "Caution",  // 경고 맥락 유지
+      "behavior": {
+        "action": "command",
+        "command": "email.resend"
+      }
+    }
+  ]
+}
+```
+
+</details>
+
+### 문제 3: 데이터 테이블 헤더
+
+```
+┌─────┬────────┬────────┬────────┐
+│ ID  │ Name   │ Email  │ Status │  ← 작은 회색 텍스트
+├─────┼────────┼────────┼────────┤
+│ 001 │ Teo    │ teo@.. │ Active │
+```
+
+<details>
+<summary>정답 보기</summary>
+
+```json
+{
+  "type": "Group",
+  "role": "Table",
+  "density": "Compact",  // 좁은 간격
+  "children": [
+    {
+      "type": "Text",
+      "role": "Label",
+      "content": "ID",
+      "prominence": "Tertiary"  // 작고 흐리게
+    },
+    {
+      "type": "Text",
+      "role": "Label",
+      "content": "Name",
+      "prominence": "Tertiary"
+    }
+    // ...
+  ]
+}
+```
+
+</details>
+
+---
+
+## 핵심 정리
+
+### IDDL의 3가지 원칙
+
+1. **의도를 선언한다** (구현을 지시하지 않는다)
+2. **4가지 질문에 답한다** (무엇? 얼마나? 무슨 색? 얼마나 빽빽하게?)
+3. **렌더러에게 맡긴다** (디자인 시스템이 알아서 렌더링)
+
+### 4가지 핵심 속성
+
+| 속성 | 질문 | 값 예시 |
+|------|------|---------|
+| `role` | 무엇인가? | Title, Form, Card |
+| `prominence` | 얼마나 눈에 띄는가? | Hero, Primary, Secondary |
+| `intent` | 무슨 맥락인가? | Neutral, Brand, Critical |
+| `density` | 얼마나 빽빽한가? | Comfortable, Standard, Compact |
+
+### 언제 어떤 값을 쓸까?
+
+```
+prominence: Hero → 페이지당 1개 (메인 타이틀, 거대 CTA)
+prominence: Primary → 주요 콘텐츠/버튼
+prominence: Secondary → 보조 정보/버튼
+prominence: Tertiary → 메타 정보, 캡션
+
+intent: Neutral → 대부분의 콘텐츠
+intent: Brand → 주요 CTA, 브랜드 강조
+intent: Positive → 성공, 완료, 긍정
+intent: Caution → 경고, 주의
+intent: Critical → 삭제, 에러, 위험
+
+density: Comfortable → 마케팅, 대시보드
+density: Standard → 일반 폼, 문서
+density: Compact → 데이터 그리드, 전문가 도구
+```
+
+---
+
+## 다음 단계
+
+축하합니다! IDDL의 핵심 철학을 이해했습니다.
+
+이제 각 속성을 **자세히** 배울 차례입니다:
+
+### Level 1: 4가지 핵심 속성
+
+1. **[Prominence 시스템](../01-fundamentals/01-prominence.md)** ← 여기서 시작!
+   - Hero, Primary, Secondary, Tertiary 완전 정복
+
+2. **[Intent 시스템](../01-fundamentals/02-intent.md)**
+   - 6가지 맥락 (Neutral, Brand, Positive, Caution, Critical, Info)
+
+3. **[Density 시스템](../01-fundamentals/03-density.md)**
+   - 3가지 밀도와 사용 시나리오
+
+4. **[Role 시스템](../01-fundamentals/04-role.md)**
+   - 노드 타입별 역할 정리
+
+---
+
+## 자주 묻는 질문
+
+### Q: 4가지 속성만으로 모든 UI를 표현할 수 있나요?
+
+대부분의 Admin UI (CRUD, 폼, 대시보드)는 가능합니다.
+차트나 맵 같은 특수 시각화는 Custom 노드가 필요할 수 있습니다.
+
+### Q: prominence와 intent를 둘 다 써야 하나요?
+
+아닙니다. 기본값이 있습니다:
+- `prominence` 생략 시 → `Primary`
+- `intent` 생략 시 → `Neutral`
+
+### Q: 디자인 시스템마다 렌더링 결과가 많이 다른가요?
+
+색상과 스타일은 다를 수 있지만, **의미**는 동일하게 전달됩니다.
+`intent: Critical`은 어떤 디자인 시스템에서든 "위험"을 의미합니다.
+
+---
+
+**이전**: [← 첫 번째 UI 만들기](./01-your-first-ui.md)
+**다음**: [Prominence 시스템 →](../01-fundamentals/01-prominence.md)
