+++ apps/docs/01-core-concepts/05-hierarchy-system.md
@@ -1,0 +1,35 @@
+# Hierarchy System
+
+> **4단계 계층 구조**: Page → Section/Overlay → Group → Item
+
+---
+
+## 개요
+
+Intent-Driven UI DSL은 **4단계의 명확한 계층 구조**를 가집니다. 각 레벨은 고유한 책임과 역할을 가지며, 이를 통해 복잡한 UI도 일관되게 표현할 수 있습니다.
+
+---
+
+## 계층 구조
+
+\`\`\`
+┌─────────────────────────────────────┐
+│ Page (페이지)                        │  ← 최상위 루트
+│  ├─ Section (섹션)                  │  ← 레이아웃 구획
+│  │   └─ Group (그룹)                │  ← 논리적 묶음
+│  │       └─ Item (아이템)           │  ← 최소 단위
+│  └─ Overlay (오버레이)              │  ← 플로팅 UI
+│      └─ Group (그룹)                │
+│          └─ Item (아이템)           │
+└─────────────────────────────────────┘
+\`\`\`
+
+---
+
+## 원칙
+
+1. **명확한 계층**: 4단계를 엄격하게 준수
+2. **역할 분리**: 각 레벨은 고유한 책임
+3. **재귀 허용**: Group은 중첩 가능
+4. **속성 상속**: 부모 속성은 자식에게 전달
+5. **오버라이드 가능**: 필요시 자식에서 속성 변경
