+++ src/apps/IDE/editor/ComponentPreview.tsx
