+++ src/apps/PPT/presentation/PresentationToolbar.tsx
