+++ src/apps/PPT/presentation/SlideCanvas.tsx
